+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2025-06-14 - 0.2.0
+
+### New features
+
+- Add command to configure formatting with Rust Analyzer and leptosfmt.
+
+### Bug fixes
+
+- Add *.vscode/* folder if not exists when configuring recommended extensions.
+- Ensure that configurations are obtained from the first workspace folder.
+
 ## 2025-06-09 - 0.1.0
 
 - Initial release